+++ docs/docs/bots/code/BlueskyBot/ValidateBotSettings.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) Barry Dorrans. All rights reserved.
+// Licensed under the MIT License.
+
+using Microsoft.Extensions.Options;
+
+namespace BlueskyBot
+{
+    [OptionsValidator]
+    public partial class ValidateBotSettings : IValidateOptions<BotSettings>
+    {
+    }
+}