+++ samples/Samples.Bot.Watcher/IFaultWatcher.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) Barry Dorrans. All rights reserved.
+// Licensed under the MIT License.
+
+namespace Samples.Bot.Watcher
+{
+    internal interface IFaultWatcher
+    {
+        public void LogFault(string fault);
+
+        public int Count { get; }
+    }
+}